remove dynamic_client.go and move functions
